correcting v2 contracts

+++ contracts/ExpiryHelperV2.sol
@@ -9,14 +9,14 @@ import {IHederaTokenService} from "./IHederaTokenServiceV2.sol";
 contract ExpiryHelper is FeeHelper {
     function createAutoRenewExpiry(
         address autoRenewAccount,
-        uint32 autoRenewPeriod
+        int64 autoRenewPeriod
     ) internal pure returns (IHederaTokenService.Expiry memory expiry) {
         expiry.autoRenewAccount = autoRenewAccount;
         expiry.autoRenewPeriod = autoRenewPeriod;
     }
 
     function createSecondExpiry(
-        uint32 second
+        int64 second
     ) internal pure returns (IHederaTokenService.Expiry memory expiry) {
         expiry.second = second;
     }
